requirements.txt
MySQL-python==1.2.5
SQLAlchemy==0.9.8
Tempita==0.5.2
argparse==1.2.1
decorator==3.4.0
pbr==0.10.0
six==1.8.0
sqlalchemy-migrate==0.9.2
sqlparse==0.1.14
wsgiref==0.1.2
